ml-services/audio-analyzer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
boto3==1.35.0
ffmpeg-python==0.2.0
numpy==1.26.4
python-multipart==0.0.12
